fitnesssite/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-admin-rangefilter==0.8.2
django-debug-toolbar==3.2.4
django-gfklookupwidget==1.0.9
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
